Add files via upload
The main objective of the project is to automate WiFi deauthentication attacks. A deauthentication attack is a type of denial-of-service (DoS) attack where an attacker sends deauthentication frames to disconnect clients from a WiFi network. This technique exploits vulnerabilities in the 802.11 WiFi protocol, which lacks encryption for management frames, making it possible to spoof these messages.